maps are taken from two sources, streaming assets folder and persistent data

+++ Assets/Scripts/MenuUI/MapSelectionMenu.cs
@@ -12,7 +12,8 @@ public class MapSelectionMenu : MonoBehaviour
 
     [Space]
     public InputField nameInput;
-    public RectTransform listContent;
+    public RectTransform listSelfCreatedContent;
+    public RectTransform listDefaultContent;
     [Space]
     public SaveLoadItem itemPrefab;
 
@@ -20,15 +21,17 @@ public class MapSelectionMenu : MonoBehaviour
 
     private void OnEnable()
     {
-        FillList();
+        Cleanup();
+
+        if (Directory.Exists(AI_File.PathSelfmadeMaps))
+            FillList(AI_File.PathSelfmadeMaps, listSelfCreatedContent);
+        if (Directory.Exists(AI_File.PathDefaultMaps))
+            FillList(AI_File.PathDefaultMaps, listDefaultContent);
     }
 
     string GetSelectedPath()
     {
-        string mapName = nameInput.text;
-        if (string.IsNullOrWhiteSpace(mapName)) return null;
-
-        string path = Path.Combine(Application.persistentDataPath, mapName + ".map");
+        string path = currentSelectedMap.mapPath;
 
         if (File.Exists(path))
             return path;
@@ -86,13 +89,13 @@ public class MapSelectionMenu : MonoBehaviour
     /// <summary>
     /// fills listContent with maps found in Application.persistentDataPath
     /// </summary>
-    void FillList()
+    void FillList(string folderPath, Transform contentTransform)
     {
         // TODO(2022-07-04) find more elegant solution
-        for (int i = 0; i < listContent.childCount; i++)
-            Destroy(listContent.GetChild(i).gameObject);
+        for (int i = 0; i < contentTransform.childCount; i++)
+            Destroy(contentTransform.GetChild(i).gameObject);
 
-        string[] paths = Directory.GetFiles(Application.persistentDataPath, "*.map");
+        string[] paths = Directory.GetFiles(folderPath, "*.map");
         Array.Sort(paths);
 
         SaveLoadItem first = null;
@@ -102,7 +105,7 @@ public class MapSelectionMenu : MonoBehaviour
             item.selectionMenu = this;
             item.MapName = Path.GetFileNameWithoutExtension(paths[i]);
             item.mapPath = paths[i];
-            item.transform.SetParent(listContent, false);
+            item.transform.SetParent(contentTransform, false);
 
             if (i == 0) first = item;
         }
@@ -117,16 +120,34 @@ public class MapSelectionMenu : MonoBehaviour
         string path = GetSelectedPath();
         if (path == null) return;
 
-        if (File.Exists(path))
-        {
-            File.Delete(path);
-        }
+        if (File.Exists(path)) File.Delete(path);
+
         nameInput.text = "";
-        FillList();
+
+        FillList(AI_File.PathSelfmadeMaps, listSelfCreatedContent);
+        FillList(AI_File.PathDefaultMaps, listDefaultContent);
     }
 
     public void BackToMainMenu()
     {
         SceneManager.LoadScene(AI_Scene.SCENENAME_MainMenu);
+    }
+
+    /// <summary>
+    /// Method to shift maps that have been created in the base folder into their own folder
+    /// </summary>
+    private void Cleanup()
+    {
+        string[] paths = Directory.GetFiles(Application.persistentDataPath, "*.map");
+
+        if (paths.Length == 0) return;
+
+        Directory.CreateDirectory(AI_File.PathSelfmadeMaps);
+
+        foreach (var path in paths)
+        {
+            string fileName = Path.GetFileName(path);
+            File.Move(path, Path.Combine(AI_File.PathSelfmadeMaps, fileName));
+        }
     }
 }
